requirements.txt
argh==0.26.2
astroid==1.6.0
attrs==17.3.0
colorama==0.3.9
decorator==4.1.2
detox==0.11
Django==2.0
django-extensions==1.9.8
django-filter==1.1.0
djangorestframework==3.7.7
docopt==0.6.2
dodgy==0.1.9
drf-writable-nested==0.3.2
enum-compat==0.0.2
eventlet==0.21.0
factory-boy==2.9.2
Faker==0.8.8
greenlet==0.4.12
ipython==6.2.1
ipython-genutils==0.2.0
isort==4.2.15
jedi==0.11.1
lazy-object-proxy==1.3.1
mccabe==0.6.1
parso==0.1.1
pathtools==0.1.2
pep8-naming==0.4.1
pexpect==4.3.1
pickleshare==0.7.4
pkg-resources==0.0.0
pluggy==0.6.0
prompt-toolkit==1.0.15
prospector==0.12.7
ptyprocess==0.5.2
py==1.5.2
pycodestyle==2.0.0
pydocstyle==2.1.1
pyflakes==1.6.0
Pygments==2.2.0
pylint==1.8.1
pylint-celery==0.3
pylint-common==0.2.5
pylint-django==0.7.2
pylint-flask==0.5
pylint-plugin-utils==0.2.6
pytest==3.3.1
pytest-django==3.1.2
pytest-watch==4.1.0
python-dateutil==2.6.1
pytz==2017.3
PyYAML==3.12
requirements-detector==0.5.2
setoptconf==0.2.0
simplegeneric==0.8.1
six==1.11.0
snowballstemmer==1.2.1
text-unidecode==1.1
tox==2.9.1
traitlets==4.3.2
typing==3.6.2
virtualenv==15.1.0
watchdog==0.8.3
wcwidth==0.1.7
wrapt==1.10.11
